% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_para.R
\name{est_para}
\alias{est_para}
\title{Estimate the parameters of LD score regression.}
\usage{
est_para(
  dat,
  M,
  two_step = T,
  idx_step1 = NULL,
  fix_intercept = F,
  n_blocks = 200,
  jackknife = T
)
}
\arguments{
\item{dat}{data}

\item{M}{number of SNPs in LD score regression}

\item{two_step}{logical}

\item{idx_step1}{vector, index of SNPs used in step 1}

\item{fix_intercept}{logical}

\item{n_blocks}{numeric}

\item{jackknife}{logical, whether to estimate the standard error}
}
\value{
slope and intercept in LD score regression
}
\description{
Estimate the parameters of LD score regression.
}
