% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ldsc.R
\name{format_ldsc}
\alias{format_ldsc}
\title{Format the LD score data.}
\usage{
format_ldsc(file_ldsc)
}
\arguments{
\item{file_ldsc}{name of the LD score file.}
}
\value{
formatted LD score data
}
\description{
Format the LD score data.
}
