% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ref.R
\name{format_ref}
\alias{format_ref}
\title{Format the reference data.}
\usage{
format_ref(X_ref, sumstats, group = NULL)
}
\arguments{
\item{X_ref}{reference genotypes data}

\item{sumstats}{summary statistics data}

\item{group}{LD block group}
}
\value{
z-score and reference data
}
\description{
Format the reference data.
}
