% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sumstats.R
\name{format_sumstats}
\alias{format_sumstats}
\title{Format the summary statistics data.}
\usage{
format_sumstats(
  file_sumstats,
  snps_merge = w_hm3.snplist,
  snps_mhc = snps_mhc,
  snp_col = "SNP",
  beta_col = "BETA",
  or_col = "OR",
  se_col = "SE",
  freq_col = "FREQ",
  a1_col = "A1",
  a2_col = "A2",
  p_col = "P",
  n_col = "N",
  ncase_col = "N_CASE",
  ncontrol_col = "N_CONTROL",
  n = NULL,
  z_col = "Z",
  info_col = "INFO",
  log_pval = F,
  chi2_max = NULL,
  min_freq = 0.05
)
}
\arguments{
\item{file_sumstats}{name of the summary statistics file}

\item{snps_merge}{data.frame with SNPs to extract}

\item{snps_mhc}{a set of SNPs that needed to be removed. For example, SNPs in MHC region.}

\item{snp_col}{name of a column with rs numbers. The default is \code{SNP}.}

\item{beta_col}{name of a column with effect sizes. The default is \code{BETA}.}

\item{or_col}{name of a column with odds ratios. The default is \code{OR}.}

\item{se_col}{name of a column with standard errors. The default is \code{SE}.}

\item{freq_col}{name of a column with effect allele frequencies. The default is \code{FREQ}.}

\item{a1_col}{name of a column with effect alleles. The default is \code{A1}.}

\item{a2_col}{name of a column with the other alleles. The default is \code{A2}.}

\item{p_col}{name of a column with p-values. The default is \code{P}.}

\item{n_col}{name of a column with sample sizes. The default is \code{N}.}

\item{ncase_col}{name of a column with the number of cases. The default is \code{N_CASE}.}

\item{ncontrol_col}{name of a column with the number of controls. The default is \code{N_CONTROL}.}

\item{n}{sample size. If the column for sample size is not available, users can use argument "n" to specify the total sample size for each SNP.}

\item{z_col}{name of a column with z scores. The default is \code{Z}.}

\item{info_col}{name of a column with imputation INFO. The default is \code{INFO}.}

\item{log_pval}{logical, whether the p-value is in -log10 scale. The default is \code{FALSE}.}

\item{chi2_max}{SNPs with tested chi^2 statistics large than chi2_max will be removed. The default is \code{80}.}

\item{min_freq}{SNPs with allele frequency less than min_freq will be removed. The default is \code{0.05}.}
}
\value{
formatted summary statistics data
}
\description{
Format the summary statistics data.
}
