% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_block.R
\name{ld_block_group}
\alias{ld_block_group}
\title{Group the SNPs using LD block data.}
\usage{
ld_block_group(ref_file, snp_col = "id", pos_col = "pos")
}
\arguments{
\item{ref_file}{name of the reference data file}

\item{snp_col}{name of a column with rs numbers. The default is \code{id}.}

\item{pos_col}{name of a column with position. The default is \code{pos}.}
}
\value{
group tibble
}
\description{
Group the SNPs using LD block data.
}
