% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpass.R
\name{xpass}
\alias{xpass}
\title{Estimate heritability using XPASS}
\usage{
xpass(z, K, n, Z = NULL)
}
\arguments{
\item{z}{z-score}

\item{K}{Kinship matrix}

\item{n}{sample size}

\item{Z}{covariates}
}
\value{
heritability
}
\description{
Estimate heritability using XPASS
}
