% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ref.R
\name{format_ref}
\alias{format_ref}
\title{Format the reference data.}
\usage{
format_ref(file_ref, sumstats)
}
\arguments{
\item{file_ref}{name of the reference data file}

\item{sumstats}{summary statistics data}
}
\value{
z-score and reference data
}
\description{
Format the reference data.
}
