% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ldsc_fit.R
\name{ldsc_fit}
\alias{ldsc_fit}
\title{Fit LD score regression.}
\usage{
ldsc_fit(
  sumstats,
  ldsc,
  two_step = T,
  fix_intercept = F,
  n_blocks = 200,
  jackknife = T
)
}
\arguments{
\item{sumstats}{summary statistics data}

\item{ldsc}{LD score data}

\item{two_step}{logical}

\item{fix_intercept}{logical}

\item{n_blocks}{numeric}

\item{jackknife}{logical, whether to estimate the standard error}
}
\value{
results of LD score regression
}
\description{
Fit LD score regression.
}
