% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpass.R
\name{xpass}
\alias{xpass}
\title{Estimate heritability using XPASS}
\usage{
xpass(z, X = NULL, K = NULL, n, Z = NULL, group = NULL)
}
\arguments{
\item{z}{z-score}

\item{X}{reference genotypes matrix}

\item{K}{Kinship matrix}

\item{n}{sample size}

\item{Z}{covariates}

\item{group}{compute standard error by LD block Jackknife}
}
\value{
heritability and standard error
}
\description{
Estimate heritability using XPASS
}
